resronsive signup & login

+++ New folder/signup/app.js
@@ -9,7 +9,7 @@ let year = document.getElementById('year')
 let email = document.getElementById('email')
 let password = document.getElementById('password')
 let gender = document.getElementsByName('gender')
-
+let existUser = false
 
 function sweetArlt(error, title, text) {
     Swal.fire({
@@ -18,6 +18,11 @@ function sweetArlt(error, title, text) {
         text: text,
     });
 }
+
+
+let usersData = JSON.parse(localStorage.getItem('users')) || []
+
+
 
 function formHandler(event) {
     event.preventDefault()
@@ -38,7 +43,7 @@ function formHandler(event) {
     }
 
     if (password.value.length < 8) {
-       return sweetArlt('error', 'Oops...', 'Password at least 8 character!')
+        return sweetArlt('error', 'Oops...', 'Password at least 8 character!')
     }
 
     let userObj = {
@@ -59,14 +64,24 @@ function formHandler(event) {
 
     }
 
-    let usersData = JSON.parse(localStorage.getItem('users')) || []
+
+    usersData.find((items) => {
+        if (items.email == email.value) {
+            existUser = true
+            return
+        }
+    })
+
+    if (existUser) {
+        return sweetArlt('error', 'Opps...', 'This email is already register')
+    }
 
     usersData.push(userObj)
 
     localStorage.setItem('users', JSON.stringify(usersData))
-    
 
-    sweetArlt('success','Signup','Congratulations! signup successfully!')
+
+    sweetArlt('success', 'Signup', 'Congratulations! signup successfully!')
 
 
     setTimeout(()=>{
@@ -76,7 +91,6 @@ function formHandler(event) {
     form.reset()
 
 }
-
 
 
 signupEl.addEventListener('click', formHandler)
